Data Science: Python Game Submission
This game is part of my Data Science: Python class. It was an exercise designed to get used to programming fundamentals and the process of coding a program on your own. 

I had a lot of fun putting together this Earthquake Adventure Game just after a 4.7 Earthquake in San Francisco. I thought it might teach one or another lesson.